# contract/openapi.yaml
...
openapi: 3.1.0
info:
  title: Certa API
  version: 0.1.0
  description: Flight operations & compliance logbook API. All quantities are SI; all instants are UTC ISO-8601. Org-scoped routes live under /api/v1/orgs/{orgId}.
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: better-auth.session_token
    bearer:
      type: http
      scheme: bearer
...
paths:
...
  /api/v1/orgs/{orgId}/rules:
    get:
      summary: Effective rule pack for the organization, with overrides applied
      tags:
        - rules
      parameters:
        - schema:
            type: string
            pattern: ^[A-Z]{2}-[A-Za-z0-9]+(-[A-Za-z0-9]+)*$
          in: query
          name: jurisdiction
          required: false
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: orgId
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  pack:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                      jurisdiction:
                        type: string
                        pattern: ^[A-Z]{2}-[A-Za-z0-9]+(-[A-Za-z0-9]+)*$
                      version:
                        type: string
                      name:
                        type: string
                      authority:
                        type: string
                      effectiveFrom:
                        type: string
                        format: date
                        pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
                      disclaimer:
                        type: string
                      ruleCount:
                        type: integer
                        minimum: -9007199254740991
                        maximum: 9007199254740991
                      unverifiedCount:
                        type: integer
                        minimum: -9007199254740991
                        maximum: 9007199254740991
                      active:
                        type: boolean
                    required:
                      - id
                      - jurisdiction
                      - version
                      - name
                      - authority
                      - effectiveFrom
                      - disclaimer
                      - ruleCount
                      - unverifiedCount
                      - active
                    additionalProperties: false
                  credentialTypes:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        label:
                          type: string
                        description:
                          type: string
                        validityRuleId:
                          type:
                            - string
                            - "null"
                        requiredForCurrency:
                          type: boolean
                      required:
                        - id
                        - label
                        - description
                        - validityRuleId
                        - requiredForCurrency
                      additionalProperties: false
                  rules:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        title:
                          type: string
                        description:
                          type: string
                        kind:
                          type: string
                        appliesTo:
                          type: string
                        value: {}
                        packValue: {}
                        unit:
                          type:
                            - string
                            - "null"
                        statedAs:
                          type:
                            - string
                            - "null"
                        source:
                          type: object
                          properties:
                            title:
                              type: string
                              minLength: 1
                            citation:
                              type: string
                              minLength: 1
                            url:
                              anyOf:
                                - type: string
                                  format: uri
                                - type: "null"
                          required:
                            - title
                            - citation
                            - url
                          additionalProperties: false
                        lastVerifiedOn:
                          anyOf:
                            - type: string
                              format: date
                              pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
                            - type: "null"
                        needsVerification:
                          type: boolean
                        notes:
                          type:
                            - string
                            - "null"
                        override:
                          anyOf:
                            - type: object
                              properties:
                                value: {}
                                reason:
                                  type: string
                                setBy:
                                  type:
                                    - string
                                    - "null"
                                setAt:
                                  type: string
                                  format: date-time
                                  pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                              required:
                                - value
                                - reason
                                - setBy
                                - setAt
                              additionalProperties: false
                            - type: "null"
                      required:
                        - id
                        - title
                        - description
                        - kind
                        - appliesTo
                        - value
                        - packValue
                        - unit
                        - statedAs
                        - source
                        - lastVerifiedOn
                        - needsVerification
                        - notes
                        - override
                      additionalProperties: false
                required:
                  - pack
                  - credentialTypes
                  - rules
                additionalProperties: false
...
security:
  - session: []
  - bearer: []